pyproject.toml
[project]
name = "beltout"
version = "0.1.0"
description = "BeltOut: An open source pitch-perfect voice-to-voice timbre transfer model based on Chatterbox"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
dependencies = [
    "numpy>=1.26.0",
    "librosa==0.11.0",
    "s3tokenizer",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchcrepe==0.0.24",
    "transformers==4.46.3",
    "diffusers==0.29.0",
    "conformer==0.3.2",
    "safetensors==0.5.3",

    # training
    "arro3-core==0.5.1",
    "arro3-io==0.5.1",
    "huggingface_hub==0.33.1",
    "scipy==1.15.3",
    # logging during training
    "tqdm==4.67.1",

    "questionary==2.1.0",
    "gradio==5.35.0",
    "soundfile==0.13.1",
]

[project.urls]
Homepage = "https://github.com/Bill13579/beltout"
Repository = "https://github.com/Bill13579/beltout"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
